requirements.txt
# Blockchain related
web3==6.0.0
eth-account==0.8.0
py-solc-x==1.1.1

# AI and Machine Learning
tensorflow==2.12.0
scikit-learn==1.2.2
opencv-python==4.7.0.72
face-recognition==1.3.0
deepface==0.0.79

# Web Framework
flask==2.3.2
flask-restful==0.3.10
flask-cors==4.0.0

# Database
pymongo==4.3.3

# Tools and Utilities
python-dotenv==1.0.0
pycryptodome==3.18.0
pyotp==2.8.0
qrcode==7.4.2
pillow==9.5.0
requests==2.31.0
numpy==1.24.3
pandas==2.0.2

# Testing
pytest==7.3.1 